RTC: the offset overflowed on every 32-bit long, and MSVC CI ran 2 of 7 tests
Answering "can't the POSIX front end run here?" turned out to be the wrong
question, and asking it found a shipped bug.

SEVEN OF THE NINE TESTS DO NOT NEED THE POSIX FRONT END.  They link
DISPATCH_OBJS - emu_io_common, hbios_dispatch, hbios_cpu, emu_init - and
nothing else; only the `test:` target's dependency on the romwbw_emu binary
made it look like the CLI was in the way.  Building them one at a time with
mingw on Windows, six passed and rtc_settim failed four checks.

`HBIOSDispatch` counted the guest's RTC offset in seconds from a fixed epoch
in a `long`.  On LP64 - Linux, macOS - that is 64 bits and it worked.  On
LLP64, which is EVERY Windows compiler, and on 32-bit Android, `long` is 32
bits and days * 86400 passes INT32_MAX in January 2038.  For the 2084 leap day
tests/rtc_settim.cc sets on purpose, 41666 * 86400 = 3,599,942,400 wrapped to
-695,024,896.  So a guest that set a date past January 2038 read back a
different one, BF_RTCSETTIM appeared to take and did not, and a second set
added to the first instead of replacing it.

z80cpmw and cpmdroid both shipped it: both compile hbios_dispatch.cc in place,
and cpmdroid builds armeabi-v7a and x86.  Measured rather than inferred - the
pre-fix file built with `cl` fails the same four checks and passes after.
rtc_offset_seconds, secondsBetween() and the civil-date conversion on both
sides are `long long` now, including the inverse, where narrowing back to
`long` would have put a silent truncation inside one round trip.

WHY IT SURVIVED, which is the more useful half.  tests/rtc_settim.cc has
always been portable - a stub CPU, no POSIX.  It had simply never been
compiled by a Windows compiler.  The msvc job ran two tests BY NAME and the
comment above that list read "two of the four tests are portable" while the
suite grew to nine, so five portable tests were running only where a 32-bit
`long` does not exist.  The job now builds and runs all seven, verified here
with cl, and the file says a new portable test belongs in that list: a test
that runs only on the POSIX job cannot see an LLP64 bug.

todo.txt: this block said "the C++ tests cannot run here" yesterday, on the
strength of the CLI not building.  Corrected, with the seven-test recipe, and
with what is actually left - cli_console needs a pty and cli_hostfile a POSIX
filesystem.  WSL Ubuntu is installed and could run both, but carries python3
and nothing else, and its sudo wants a password this session cannot give.

+++ src/hbios_dispatch.cc
@@ -1702,42 +1702,52 @@ void HBIOSDispatch::handleDIO() {
 // rather than mktime(), which is local-time and DST-dependent: what is wanted
 // here is the difference between two calendar readings, not between two
 // instants in somebody's timezone.
-static long days_from_civil(int y, int m, int d) {
+static long long days_from_civil(int y, int m, int d) {
   y -= m <= 2;
-  const long era = (y >= 0 ? y : y - 399) / 400;
+  const long long era = (y >= 0 ? y : y - 399) / 400;
   const unsigned yoe = (unsigned)(y - era * 400);
   const unsigned doy = (unsigned)((153 * (m + (m > 2 ? -3 : 9)) + 2) / 5 + d - 1);
   const unsigned doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
-  return era * 146097L + (long)doe - 719468L;
+  return era * 146097LL + (long long)doe - 719468LL;
 }
 
-static long civil_seconds(const emu_time& t) {
-  return days_from_civil(t.year, t.month, t.day) * 86400L
-       + (long)t.hour * 3600L + (long)t.minute * 60L + (long)t.second;
+// long long throughout: on LLP64 (every Windows compiler) and on 32-bit
+// Android, `long` is 32 bits, and days * 86400 passes INT32_MAX in
+// January 2038 - 41666 * 86400 for the 2084 leap day this file used to
+// compute as -695024896.  The arithmetic below is deliberately not
+// mktime(), so it does not inherit a 64-bit time_t from the host either;
+// the width has to be stated here.
+static long long civil_seconds(const emu_time& t) {
+  return days_from_civil(t.year, t.month, t.day) * 86400LL
+       + (long long)t.hour * 3600LL + (long long)t.minute * 60LL
+       + (long long)t.second;
 }
 
-long HBIOSDispatch::secondsBetween(const emu_time& a, const emu_time& b) {
+long long HBIOSDispatch::secondsBetween(const emu_time& a, const emu_time& b) {
   return civil_seconds(b) - civil_seconds(a);
 }
 
 void HBIOSDispatch::applyRtcOffset(emu_time* t) const {
   if (!t || rtc_offset_seconds == 0) return;
 
-  long secs = civil_seconds(*t) + rtc_offset_seconds;
-  long days = secs / 86400L;
-  long rem = secs % 86400L;
-  if (rem < 0) { rem += 86400L; days -= 1; }
+  long long secs = civil_seconds(*t) + rtc_offset_seconds;
+  long long days = secs / 86400LL;
+  long long rem = secs % 86400LL;
+  if (rem < 0) { rem += 86400LL; days -= 1; }
 
   t->hour = (int)(rem / 3600);
   t->minute = (int)((rem % 3600) / 60);
   t->second = (int)(rem % 60);
 
-  // civil_from_days, the inverse of the above.
-  long z = days + 719468L;
-  const long era = (z >= 0 ? z : z - 146096) / 146097;
+  // civil_from_days, the inverse of the above.  Same width as `days` on
+  // purpose: a day count fits 32 bits for any date anyone will type, but
+  // narrowing it back here would put a silent truncation between the two
+  // halves of one round trip.
+  long long z = days + 719468LL;
+  const long long era = (z >= 0 ? z : z - 146096) / 146097;
   const unsigned doe = (unsigned)(z - era * 146097);
   const unsigned yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
-  long y = (long)yoe + era * 400;
+  long long y = (long long)yoe + era * 400;
   const unsigned doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
   const unsigned mp = (5 * doy + 2) / 153;
   const unsigned d = doy - (153 * mp + 2) / 5 + 1;
@@ -1746,7 +1756,7 @@ void HBIOSDispatch::applyRtcOffset(emu_time* t) const {
   t->month = (int)m;
   t->day = (int)d;
   // 1970-01-01 was a Thursday.
-  long wd = (days + 4) % 7;
+  long long wd = (days + 4) % 7;
   if (wd < 0) wd += 7;
   t->weekday = (int)wd;
 }
